Add management of messages from containers
Add a way to map the PID of a containerized process to its PID on the
host machine. It is now possible to receive add and feedback messages
directly from containers. Containers wrapped inside Kubernetes pods are
also supported. The HTTP Server automatically detects the type of the
application by inspecting the kernel namespace pseudo-fs.